VM: Add BaseStateManager (#1548)
* vm -> BaseStateManager: added class and constructor, improved DefaultStateManager class documentation

* VM -> StateManager: removed unused warm() method from Cache

* VM -> StateManager: made Cache backend agnostic by replacing direct trie calls with operation callbacks

* VM -> StateManager: moved getAccount(), putAccount() and touchAccount() to BaseStateManager, made _debug a member variable of BaseStateManater

* VM -> StateManager: moved deleteAccount(), getOriginalContractStorage(), clearOriginalStorageCache(), generateAccessList() to BaseStateManager, checkpoint(), commit(), revert() calls into super class

* VM -> StateManager: use putAccount() in generateGenesis() in favor of direct trie access, move generateCanonicalGenesis(), generateGenesis() to BaseStateManager()

* VM -> StateManager: moved generateGenesis() back to StateManager (direct trie access needed), moved EIP-2929 functionality (addWarmedAddress(),...) to BaseStateManager

* VM -> StateManager: some cleanup

* VM -> StateManager: flush cache in generateGenesis() call to allow for directly using the provided putAccount() method

* VM -> StateManager: moved generateGenesis() to BaseStateManager
