+++ packages/vm/src/state/index.ts
@@ -1,2 +1,3 @@
 export { StateManager } from './interface'
+export { BaseStateManager } from './baseStateManager'
 export { default as DefaultStateManager } from './stateManager'
